+++ app/views/additional_payments/claims/poor_performance.html.erb
@@ -1,89 +1,31 @@
-<% content_for(:page_title, page_title(t("#{@form.i18n_namespace}.forms.poor_performance.questions.poor_performance"), journey: current_journey_routing_name, show_error: @form.errors.any?)) %>
+<% content_for(:page_title, page_title(@form.t(:heading), journey: current_journey_routing_name, show_error: @form.errors.any?)) %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= render("shared/error_summary", instance: @form, errored_field_id_overrides: {
-      "subject_to_formal_performance_action": "claim_subject_to_formal_performance_action_true",
-      "subject_to_disciplinary_action": "claim_subject_to_disciplinary_action_true"
-    }) if @form.errors.any? %>
-    <%= form_for @form, url: claim_path(current_journey_routing_name) do |form| %>
+    <%= form_for @form, url: claim_path(current_journey_routing_name), builder: GOVUKDesignSystemFormBuilder::FormBuilder do |f| %>
+      <%= f.govuk_error_summary %>
+
       <h1 class="govuk-heading-l">
-        <%= t("#{@form.i18n_namespace}.forms.poor_performance.questions.poor_performance") %>
+        <%= @form.t(:heading) %>
       </h1>
-      <%= form_group_tag @form, :subject_to_formal_performance_action do %>
 
-        <%= form.hidden_field :subject_to_formal_performance_action %>
+      <%= f.govuk_collection_radio_buttons :subject_to_formal_performance_action, @form.radio_options, :id, :name,
+        legend: {
+          text: @form.t("questions.performance.question"),
+          tag: "h2",
+          size: "m"
+        },
+        hint: { text: @form.t("questions.performance.hint") } %>
 
-        <fieldset class="govuk-fieldset" aria-describedby="formal_performance_action-hint">
+      <%= f.govuk_collection_radio_buttons :subject_to_disciplinary_action, @form.radio_options, :id, :name,
+        legend: {
+          text: @form.t("questions.disciplinary.question"),
+          tag: "h2",
+          size: "m"
+        },
+        hint: { text: @form.t("questions.disciplinary.hint") } %>
 
-          <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
-            <h2 class="govuk-fieldset__heading">
-              <%= t("#{@form.i18n_namespace}.forms.poor_performance.questions.formal_performance_action") %>
-            </h2>
-          </legend>
-
-          <%= errors_tag @form, :subject_to_formal_performance_action %>
-
-          <div id="formal_performance_action-hint" class="govuk-hint">
-            <%= t("#{@form.i18n_namespace}.forms.poor_performance.questions.formal_performance_action_hint") %>
-          </div>
-
-          <div class="govuk-radios">
-
-            <div class="govuk-radios__item">
-              <%= form.radio_button(:subject_to_formal_performance_action, true, class: "govuk-radios__input") %>
-              <%= form.label :subject_to_formal_performance_action_true, "Yes", class: "govuk-label govuk-radios__label" %>
-            </div>
-
-            <div class="govuk-radios__item">
-              <%= form.radio_button(:subject_to_formal_performance_action, false, class: "govuk-radios__input") %>
-              <%= form.label :subject_to_formal_performance_action_false, "No", class: "govuk-label govuk-radios__label" %>
-            </div>
-
-          </div>
-
-        </fieldset>
-
-      <% end %>
-
-      <%= form_group_tag @form, :subject_to_disciplinary_action do %>
-
-        <%= form.hidden_field :subject_to_disciplinary_action %>
-
-        <fieldset class="govuk-fieldset" aria-describedby="disciplinary_action-hint">
-
-          <legend class="govuk-fieldset__legend govuk-fieldset__legend--m">
-            <h2 class="govuk-fieldset__heading">
-              <%= t("#{@form.i18n_namespace}.forms.poor_performance.questions.disciplinary_action") %>
-            </h2>
-          </legend>
-
-          <%= errors_tag @form, :subject_to_disciplinary_action %>
-
-          <div id="disciplinary_action-hint" class="govuk-hint">
-            <%= t("#{@form.i18n_namespace}.forms.poor_performance.questions.disciplinary_action_hint") %>
-          </div>
-
-          <div class="govuk-radios">
-
-            <div class="govuk-radios__item">
-              <%= form.radio_button(:subject_to_disciplinary_action, true, class: "govuk-radios__input") %>
-              <%= form.label :subject_to_disciplinary_action_true, "Yes", class: "govuk-label govuk-radios__label" %>
-            </div>
-
-            <div class="govuk-radios__item">
-              <%= form.radio_button(:subject_to_disciplinary_action, false, class: "govuk-radios__input") %>
-              <%= form.label :subject_to_disciplinary_action_false, "No", class: "govuk-label govuk-radios__label" %>
-            </div>
-
-          </div>
-
-        </fieldset>
-
-      <% end %>
-
-
-      <%= form.submit "Continue", class: "govuk-button", data: {module: "govuk-button"} %>
+      <%= f.govuk_submit %>
     <% end %>
   </div>
 </div>
