+++ spec/features/backlink_spec.rb
@@ -24,8 +24,12 @@ RSpec.feature "Backlinking during a claim" do
     click_on "Continue"
 
     expect(page).to have_content("Tell us if you are currently under any performance measures or disciplinary action")
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     expect(page).to have_content("Which route into teaching did you take?")
