+++ spec/features/combined_teacher_claim_journey_with_teacher_id_check_mobile_spec.rb
@@ -132,8 +132,12 @@ RSpec.feature "Combined journey with Teacher ID mobile check" do
     click_on "Continue"
 
     # - Performance Issues
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     # - Check and confirm qualifications
