+++ spec/features/combined_teacher_claim_journey_with_teacher_id_spec.rb
@@ -362,10 +362,14 @@ RSpec.feature "Combined journey with Teacher ID" do
     click_on "Continue"
 
     # - Performance Issues
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.poor_performance"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.heading"))
 
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
   end
 end
