+++ spec/features/early_career_payments_claim_lupp_ineligible_spec.rb
@@ -45,18 +45,20 @@ RSpec.feature "Early-Career Payments claims with school ineligible for Levelling
     click_on "Continue"
 
     # - Performance Issues
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.poor_performance"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action_hint"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.heading"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.hint"))
 
-    # No
-    choose "claim_subject_to_formal_performance_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
 
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action_hint"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.hint"))
 
-    # "No"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
 
     click_on "Continue"
