+++ spec/features/ineligible_early_career_payments_claims_spec.rb
@@ -70,11 +70,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_true"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("Yes")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     expect(page).to have_text(I18n.t("additional_payments.ineligible.heading"))
@@ -110,11 +114,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_true"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("Yes")
+    end
     click_on "Continue"
 
     expect(page).to have_text(I18n.t("additional_payments.ineligible.heading"))
@@ -150,11 +158,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_true"
-    choose "claim_subject_to_disciplinary_action_true"
+    within all(".govuk-fieldset")[0] do
+      choose("Yes")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("Yes")
+    end
     click_on "Continue"
 
     expect(page).to have_text(I18n.t("additional_payments.ineligible.heading"))
@@ -276,11 +288,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     # - What route into teaching did you take?
@@ -341,11 +357,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     # - What route into teaching did you take?
@@ -413,11 +433,15 @@ RSpec.feature "Ineligible Teacher Early-Career Payments claims", slow: true do
     click_on "Continue"
 
     # - Poor performance
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.formal_performance_action"))
-    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary_action"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.performance.question"))
+    expect(page).to have_text(I18n.t("additional_payments.forms.poor_performance.questions.disciplinary.question"))
 
-    choose "claim_subject_to_formal_performance_action_false"
-    choose "claim_subject_to_disciplinary_action_false"
+    within all(".govuk-fieldset")[0] do
+      choose("No")
+    end
+    within all(".govuk-fieldset")[1] do
+      choose("No")
+    end
     click_on "Continue"
 
     # - What route into teaching did you take?
